+++ src/Organizations/OrganizationEdit/OrganizationEdit.js
@@ -1,37 +1,40 @@
+import PropTypes from 'prop-types';
 import {
   useCallback,
-  useEffect,
   useMemo,
-  useState,
 } from 'react';
-import { FormattedMessage, useIntl } from 'react-intl';
+import {
+  FormattedMessage,
+  useIntl,
+} from 'react-intl';
 import { withRouter } from 'react-router-dom';
-import PropTypes from 'prop-types';
 
-import {
-  stripesConnect,
-} from '@folio/stripes/core';
-import {
-  Paneset,
-} from '@folio/stripes/components';
+import { Paneset } from '@folio/stripes/components';
 import {
   LoadingPane,
+  useOrganization,
   useShowCallout,
 } from '@folio/stripes-acq-components';
 
-import { VIEW_ORG_DETAILS } from '../../common/constants';
-import { useOrganizationBankingInformation } from '../../common/hooks';
-import { organizationResourceByUrl } from '../../common/resources';
-import { BANKING_INFORMATION_FIELD_NAME } from '../constants';
+import {
+  ORGANIZATIONS_ROUTE,
+  VIEW_ORG_DETAILS,
+} from '../../common/constants';
+import {
+  useOrganizationBankingInformation,
+  useOrganizationMutation,
+} from '../../common/hooks';
+import {
+  BANKING_INFORMATION_FIELD_NAME,
+  SUBMIT_ACTION,
+  SUBMIT_ACTION_FIELD_NAME,
+} from '../constants';
 import { OrganizationForm } from '../OrganizationForm';
 import { handleSaveErrorResponse } from '../handleSaveErrorResponse';
 import { useBankingInformationManager } from '../useBankingInformationManager';
 
-export const OrganizationEdit = ({ match, history, location, mutator }) => {
+export const OrganizationEdit = ({ match, history, location }) => {
   const organizationId = match.params.id;
-
-  const [organization, setOrganization] = useState({});
-  const [isOrganizationLoading, setIsOrganizationLoading] = useState(true);
   const showCallout = useShowCallout();
   const intl = useIntl();
 
@@ -42,17 +45,26 @@ export const OrganizationEdit = ({ match, history, location, mutator }) => {
     isLoading: isBankingInformationLoading,
   } = useOrganizationBankingInformation(organizationId);
 
-  useEffect(
-    () => {
-      mutator.editOrganizationOrg.GET()
-        .then(organizationsResponse => {
-          setOrganization(organizationsResponse);
-        })
-        .finally(() => setIsOrganizationLoading(false));
-    },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [],
-  );
+  const {
+    organization,
+    isFetching: isOrganizationFetching,
+    isLoading: isOrganizationLoading,
+    refetch,
+  } = useOrganization(organizationId);
+
+  const {
+    updateOrganization,
+    isLoading: isUpdateOrganizationLoading,
+  } = useOrganizationMutation();
+
+  const saveAndKeepEditingHandler = useCallback(async () => {
+    await refetch();
+
+    history.push({
+      pathname: `${ORGANIZATIONS_ROUTE}/${organizationId}/edit`,
+      search: location.search,
+    });
+  }, [history, location.search, organizationId, refetch]);
 
   const cancelForm = useCallback(
     () => {
@@ -62,36 +74,45 @@ export const OrganizationEdit = ({ match, history, location, mutator }) => {
         state: { isDetailsPaneInFocus: true },
       });
     },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [location.search],
+    [history, location.search, organizationId],
   );
 
-  const updateOrganization = useCallback(
-    (values, { getFieldState }) => {
-      const { [BANKING_INFORMATION_FIELD_NAME]: bankingInformation, ...data } = values;
+  const onSubmit = useCallback((
+    { [SUBMIT_ACTION_FIELD_NAME]: submitAction, ...values },
+    form,
+  ) => {
+    const { [BANKING_INFORMATION_FIELD_NAME]: bankingInformation, ...data } = values;
 
-      return mutator.editOrganizationOrg.PUT(data)
-        .then(() => {
-          return manageBankingInformation({
-            initBankingInformation: getFieldState(BANKING_INFORMATION_FIELD_NAME)?.initial,
-            bankingInformation,
-            organization: values,
-          });
-        })
-        .then(() => {
-          setTimeout(cancelForm);
-          showCallout({
-            messageId: 'ui-organizations.save.success',
-            values: { organizationName: data.name },
-          });
-        })
-        .catch(async e => {
-          await handleSaveErrorResponse(intl, showCallout, e);
+    return updateOrganization({ data })
+      .then(() => {
+        return manageBankingInformation({
+          initBankingInformation: form.getFieldState(BANKING_INFORMATION_FIELD_NAME)?.initial,
+          bankingInformation,
+          organization: values,
         });
-    },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [cancelForm, intl, manageBankingInformation, showCallout],
-  );
+      })
+      .then(async () => {
+        showCallout({
+          messageId: 'ui-organizations.save.success',
+          values: { organizationName: data.name },
+        });
+
+        switch (submitAction) {
+          case SUBMIT_ACTION.saveAndKeepEditing:
+            await saveAndKeepEditingHandler();
+            form.restart();
+            break;
+          case SUBMIT_ACTION.saveAndClose:
+          default:
+            cancelForm();
+            break;
+        }
+      })
+      .catch(async (e) => {
+        await handleSaveErrorResponse(intl, showCallout, e?.response);
+      });
+  },
+  [cancelForm, intl, manageBankingInformation, saveAndKeepEditingHandler, showCallout, updateOrganization]);
 
   const initialValues = useMemo(() => ({
     [BANKING_INFORMATION_FIELD_NAME]: bankingInformationData,
@@ -111,25 +132,18 @@ export const OrganizationEdit = ({ match, history, location, mutator }) => {
   return (
     <OrganizationForm
       initialValues={initialValues}
-      onSubmit={updateOrganization}
+      isSubmitDisabled={isOrganizationFetching || isUpdateOrganizationLoading}
+      onSubmit={onSubmit}
       cancelForm={cancelForm}
       paneTitle={<FormattedMessage id="ui-organizations.editOrg.title" values={{ name: organization.name }} />}
     />
   );
 };
 
-OrganizationEdit.manifest = Object.freeze({
-  editOrganizationOrg: {
-    ...organizationResourceByUrl,
-    accumulate: true,
-  },
-});
-
 OrganizationEdit.propTypes = {
   match: PropTypes.object.isRequired,
   history: PropTypes.object.isRequired,
   location: PropTypes.object.isRequired,
-  mutator: PropTypes.object.isRequired,
 };
 
-export default withRouter(stripesConnect(OrganizationEdit));
+export default withRouter(OrganizationEdit);