+++ src/Organizations/OrganizationEdit/OrganizationEdit.test.js
@@ -1,15 +1,34 @@
-import { QueryClient, QueryClientProvider } from 'react-query';
+import {
+  QueryClient,
+  QueryClientProvider,
+} from 'react-query';
 
-import { render, screen } from '@folio/jest-config-stripes/testing-library/react';
+import {
+  render,
+  screen,
+} from '@folio/jest-config-stripes/testing-library/react';
+import { useOrganization } from '@folio/stripes-acq-components';
 
-import { useOrganizationBankingInformation } from '../../common/hooks';
+import {
+  useOrganizationBankingInformation,
+  useOrganizationMutation,
+} from '../../common/hooks';
+import {
+  SUBMIT_ACTION,
+  SUBMIT_ACTION_FIELD_NAME,
+} from '../constants';
 import { OrganizationForm } from '../OrganizationForm';
 import { useBankingInformationManager } from '../useBankingInformationManager';
 import { OrganizationEdit } from './OrganizationEdit';
 
+jest.mock('@folio/stripes-acq-components', () => ({
+  ...jest.requireActual('@folio/stripes-acq-components'),
+  useOrganization: jest.fn(),
+}));
 jest.mock('../../common/hooks', () => ({
   ...jest.requireActual('../../common/hooks'),
   useOrganizationBankingInformation: jest.fn(),
+  useOrganizationMutation: jest.fn(),
 }));
 jest.mock('../OrganizationForm', () => ({
   OrganizationForm: jest.fn().mockReturnValue('OrganizationForm'),
@@ -29,12 +48,6 @@ const bankingInformation = [{
   isPrimary: true,
 }];
 
-const mutatorMock = {
-  editOrganizationOrg: {
-    GET: jest.fn(),
-    PUT: jest.fn(),
-  },
-};
 const historyMock = {
   push: jest.fn(),
 };
@@ -46,21 +59,25 @@ const matchMock = {
 
 const getFieldState = jest.fn();
 const manageBankingInformation = jest.fn();
+const refetchOrganization = jest.fn(() => Promise.resolve());
+const updateOrganization = jest.fn(() => Promise.resolve());
 
 const queryClient = new QueryClient();
-
 const wrapper = ({ children }) => (
   <QueryClientProvider client={queryClient}>
     {children}
   </QueryClientProvider>
 );
 
-const renderOrganizationEdit = (props) => render(
+const defaultProps = {
+  history: historyMock,
+  match: matchMock,
+  location: {},
+};
+
+const renderOrganizationEdit = (props = {}) => render(
   <OrganizationEdit
-    match={matchMock}
-    location={{}}
-    history={historyMock}
-    mutator={mutatorMock}
+    {...defaultProps}
     {...props}
   />,
   { wrapper },
@@ -72,15 +89,19 @@ describe('OrganizationEdit', () => {
 
     getFieldState.mockClear();
     historyMock.push.mockClear();
-    mutatorMock.editOrganizationOrg.GET.mockClear().mockReturnValue(Promise.resolve(organization));
-    mutatorMock.editOrganizationOrg.PUT.mockClear();
+    useOrganization.mockClear().mockReturnValue({
+      organization,
+      isFetching: false,
+      refetch: refetchOrganization,
+    });
 
-    useOrganizationBankingInformation
-      .mockClear()
-      .mockReturnValue({ bankingInformation, isLoading: false });
-    useBankingInformationManager
-      .mockClear()
-      .mockReturnValue({ manageBankingInformation });
+    useBankingInformationManager.mockReturnValue({ manageBankingInformation });
+    useOrganizationBankingInformation.mockReturnValue({ bankingInformation, isLoading: false });
+    useOrganizationMutation.mockReturnValue({ updateOrganization });
+  });
+
+  afterEach(() => {
+    jest.clearAllMocks();
   });
 
   it('should display organization form', async () => {
@@ -102,24 +123,34 @@ describe('OrganizationEdit', () => {
   });
 
   it('should save organization', async () => {
-    mutatorMock.editOrganizationOrg.PUT.mockReturnValue(Promise.resolve({ id: 'orgUid' }));
-
     renderOrganizationEdit();
 
     await screen.findByText('OrganizationForm');
     await OrganizationForm.mock.calls[0][0].onSubmit({}, { getFieldState });
 
-    expect(mutatorMock.editOrganizationOrg.PUT).toHaveBeenCalled();
+    expect(updateOrganization).toHaveBeenCalled();
   });
 
   it('should handle banking information on form submit', async () => {
-    mutatorMock.editOrganizationOrg.PUT.mockReturnValue(Promise.resolve({ id: 'orgUid' }));
-
     renderOrganizationEdit();
 
     await screen.findByText('OrganizationForm');
     await OrganizationForm.mock.calls[0][0].onSubmit({}, { getFieldState });
 
     expect(manageBankingInformation).toHaveBeenCalled();
+  });
+
+  it('should restart and refetch data on saveAndKeepEditing', async () => {
+    const restart = jest.fn();
+
+    renderOrganizationEdit();
+
+    await screen.findByText('OrganizationForm');
+    await OrganizationForm.mock.calls[0][0].onSubmit({
+      [SUBMIT_ACTION_FIELD_NAME]: SUBMIT_ACTION.saveAndKeepEditing,
+    }, { getFieldState, restart });
+
+    expect(refetchOrganization).toHaveBeenCalled();
+    expect(restart).toHaveBeenCalled();
   });
 });